backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.3
pydantic==2.11.7
python-multipart==0.0.20
supabase==2.15.3
python-dotenv==1.1.0
httpx==0.28.1
google-generativeai==0.8.5
google-cloud-aiplatform==1.97.0
google-api-python-client==2.173.0
google-auth-oauthlib==1.2.2
autopep8==2.3.2